Peform a differential expression analysis.

Use fdge_model_def() to define the design matrix and contrast to test and pass the FacileDgeModelDefinition object returned from that to fdge() to run the desired differential testing framework (dictated by the method parameter) over the data. fdge_model_def accepts a

# S3 method for FacileTtestAnalysisResult
compare(x, y, ...)

# S3 method for FacileTtestAnalysisResult
report(x, type = c("dge",
  "features"), ntop = 200, max_padj = 0.1, min_logFC = 1,
  features = NULL, round_digits = 3, event_source = "A",
  webgl = TRUE, caption = NULL, ...)

fdge(x, ...)

# S3 method for FacileAnovaModelDefinition
fdge(x, assay_name = NULL,
  method = NULL, filter = "default", with_sample_weights = FALSE,
  ...)

# S3 method for FacileDgeModelDefinition
fdge(x, assay_name = NULL,
  method = NULL, filter = "default", with_sample_weights = FALSE,
  treat_lfc = NULL, ...)

Arguments

x
a data source

...
passed down into inner methods, such as biocbox to tweak filtering criteria, for instance

assay_name
the name of the assay that holds the measurements for test. Defaults to default_assay(x).

method
The differential testing framework to use over the data. The valid choices are defined by the type of assay assay_nameis. Refer to the Differential Expression Testing Methods section for more details.

filter, with_sample_weights
Passed into biocbox() to determine which features (genes) are removed from the dataset for testing, as well as if to use limma::arrayWeights() or limma::voomWithQualityWeights() (where appropriate) when testing (default is not to).

Comapring DGE Results

We can compare two Ttest results

Interacting with results

The report function will create an htmlwidget which can be explored by the analyst or dropped into an Rmarkdown report.

report(result, "dge", max_padj = 0.05, min_logFC = 1) will create a side-by-side volcano and datatable for differential expression results.

Differential Expression Testing Methods

The appropriate statistical framework to use for differential expression testing is defined by the type of data that is recorded in the assay assay_name, ie. assay_info(x, assay_name)$assay_type.

The fdge_methods() function returns a tibble of appropriate assay_type -> dge_method associations. The first entry for each dge_method is the default method used if one isn't provided by the caller.

The available methods are:

"voom": For count data, uses limma::voomWithQualityWeights() when with_sample_weights = TRUE.

"edgeR-qlf": The edgeR quasi-likelihood method, for count data.

"limma-trend": Usable for log-transformed data that "looks like" it came from count data, or where there is a "trend" of the variance with the mean, uses lima::arrayWeights() when with_sample_weights = TRUE.

"limma": Straightup limma, this expects log2-normal like data, with (largely) no trend of the variance to the mean worth modeling. Uses lima::arrayWeights() when with_sample_weights = TRUE

Examples

# Comparing two T-test results ----------------------------------------------
# Let's compare the tumor vs normal DGE results in CRC vs BLCA

efds <- exampleFacileDataSet()

#> Error in exampleFacileDataSet(): could not find function "exampleFacileDataSet"

dge.crc <- filter_samples(efds, indication == "CRC") %>%
  fdge_model_def("sample_type", "tumor", "normal", "sex") %>%
  fdge()

#> Error in filter_samples(efds, indication == "CRC"): could not find function "filter_samples"

dge.blca <- filter_samples(efds, indication == "BLCA") %>%
  fdge_model_def("sample_type", "tumor", "normal", "sex") %>%
  fdge()

#> Error in filter_samples(efds, indication == "BLCA"): could not find function "filter_samples"

dge.comp <- compare(dge.crc, dge.blca)

#> Error in compare(dge.crc, dge.blca): object 'dge.crc' not found

if (interactive()) {
  report(dge.comp)
  shine(dge.comp)
}
efds <- FacileData::exampleFacileDataSet()
samples <- FacileData::filter_samples(efds, indication == "BLCA")
mdef <- fdge_model_def(samples, covariate = "sample_type",
                       numer = "tumor", denom = "normal", fixed = "sex")
dge <- fdge(mdef, method = "voom")
dge.sig <- signature(dge)

#> Error in UseMethod("signature", x): no applicable method for 'signature' applied to an object of class "c('Facile%sFeatureRanks%s', 'FacileFeatureRanks', 'FacileAnalysisResult')"

if (interactive()) {
  viz(dge)
  shine(dge)
}

stage.anova <- samples %>%
  fdge_model_def(covariate = "stage", fixed = "sex") %>%
  fdge(method = "voom")
anova.sig <- signature(stage.anova)

#> Error in UseMethod("signature", x): no applicable method for 'signature' applied to an object of class "c('Facile%sFeatureRanks%s', 'FacileFeatureRanks', 'FacileAnalysisResult')"