Fully interactive differential expression analysis

Assembles a shiny UI to define all of the bits required to perform a differential expression analysis over a predefined set of samples.

fdgeGadget(x, title = "Differential Expression Analysis", height = 800,
  width = 1000, viewer = "browser", ...)

fdgeAnalysis(input, output, session, rfds, ..., debug = FALSE)

Arguments

model_def
the module returned from fdgeModelDefModule()

Value

a ReactiveFacileDgeAnalysisResult, the output from fdge()

Details

An interactive differential expression analysis is divided into three steps, each of which provides its own shiny module and interface. The minimal input to this analysis is a pre-defined subset of samples to act on.

These steps are:

Model matrix definition. The functionality is provided by the flm_def() function, and the shiny interface by the flmDefRun() module.

Differential expression analysis. The functionality is defined by the fdge() function, and the shiny interface by the fdgeRun() module.

Results display. The interactive display of the results is provided by the fdgeView() module.

Th

Interactive Gadget

Analysis Module

Wrapper module to perform and interact with a differential expression result.

This module can be embedded within a shiny app, or called from a gadget.

Examples

if (interactive()) {
# run tumor vs normal comparisons vs each, then run compare() on the results
options(facile.log.level.fshine = "trace")
efds <- FacileData::exampleFacileDataSet()
dge.crc <- efds %>%
  FacileData::filter_samples(indication == "CRC") %>%
  fdgeGadget(viewer = "pane")
dge.blca <- efds %>%
  filter_samples(indication == "BLCA") %>%
  fdgeGadget(viewer = "pane")
dge.comp <- compare(dge.crc, dge.blca)

if (FALSE) {
tfds <- FacileDataSet("~/workspace/data/FacileData/dockerlink/FacileTcgaDataSet")
tsamples <- filter_samples(tfds, indication == "BRCA")
tdge <- fdgeGadget(tsamples, viewer = "browser")
}
report(dge.comp)
shine(dge.comp)
}