Full interactive executation of GSEA, soup to nuts.

For now, feature set enrichment analysis are only performed downstream of a FacileAnalysisResult. This requires that we have an ffsea method defined for the specific class of the result, ie. ffsea.FacilePcaAnalysisResult.

ffseaGadget(
  x,
  gdb,
  title = "Feature Set Enrichment Analysis",
  height = 800,
  width = 1000,
  viewer = "browser",
  ...,
  debug = FALSE
)

ffseaRun(input, output, session, rfds, aresult, gdb, ..., debug = FALSE)

Arguments

x
A FacileAnalysisResult that has an implemented ffsea.* method

gdb
A reactive(GeneSetDb) object

aresult
A FacileAnalysisResult that has a ffsea.* method defined.

Custom Run Configuration

The options presented to the user for running a feature set enrichment analysis on a FacilePcaAnalysisResult will be different than the ones made available for an enrichment analysis over a FacileTtestAnalysisResult or even a FacileAnovaAnalysisResult.

As such, each type of result should define a UI that accepts the appropriate parameters for its corresponding ffsea.* method, and a server function that extract and invokes the function.

Examples

gdb <- sparrow::exampleGeneSetDb()
dge.crc <- FacileData::exampleFacileDataSet() %>%
  FacileData::filter_samples(indication == "CRC") %>%
  flm_def(covariate = "sample_type", numer = "tumor", denom = "normal",
          batch = "sex") %>%
  fdge(method = "voom")
if (interactive()) {
  fres <- ffseaGadget(dge.crc, gdb)
}