Converts an analysis result into a "signature" type of thing.

This is mainly for feature rankings, but could be something else? Returns a table of features that can be piped into GeneSetDb() constructor, should insipiration strike.

signature(x, ...)

Arguments

x

A FacileAnalysisResult.

Details

Most of the signature.* functions defined over the result of analysis first passes through a call to ranks() first, which then calls the signature() method on those ranks, ie. the two segments below will produce the same result.

sigs1 <- FacileData::exampleFacileDataSet() %>%
  FacileData::filter_samples(indication == "CRC") %>%
  fpca() %>%
  signature()
sigs2 <- FacileData::exampleFacileDataSet() %>%
  FacileData::filter_samples(indication == "CRC") %>%
  fpca() %>%
  ranks() %>%
  signature()

Examples

pca.sigs <- FacileData::exampleFacileDataSet() %>%
  FacileData::filter_samples(indication == "CRC") %>%
  fpca() %>%
  signature()